requirements.txt
streamlit
openai-whisper
torch
torchvision
transformers
pillow
opencv-python-headless
sentence-transformers
numpy